# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountRole": {
        "enum": [
          "user",
          "super_admin"
        ],
        "title": "AccountRole",
        "type": "string"
      },
      "AccountTier": {
        "enum": [
          "free",
          "pro"
        ],
        "title": "AccountTier",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExchangeRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[A-Za-z0-9_-]{43}$",
            "title": "Code",
            "type": "string"
          },
          "code_verifier": {
            "pattern": "^[A-Za-z0-9._~-]{43,128}$",
            "title": "Code Verifier",
            "type": "string"
          }
        },
        "required": [
          "code",
          "code_verifier"
        ],
        "title": "ExchangeRequest",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/Provider"
            },
            "title": "Providers",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/AccountRole"
          },
          "tier": {
            "$ref": "#/components/schemas/AccountTier"
          }
        },
        "required": [
          "id",
          "display_name",
          "providers",
          "tier",
          "role"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "Provider": {
        "enum": [
          "naver",
          "kakao"
        ],
        "title": "Provider",
        "type": "string"
      },
      "ProviderAvailability": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "provider": {
            "$ref": "#/components/schemas/Provider"
          }
        },
        "required": [
          "provider",
          "enabled"
        ],
        "title": "ProviderAvailability",
        "type": "object"
      },
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "format": "password",
            "title": "Refresh Token",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_expires_in": {
            "title": "Refresh Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "refresh_token",
          "refresh_expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "네이버·카카오 로그인 후 발급한 Mori 액세스 토큰",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "소셜 가입, 최고 관리자 승인 기반 프로 권한, 사용자별 주차 기록 저장·조회",
    "title": "Mori API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/exchange": {
      "post": {
        "operationId": "exchange_v1_auth_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "일회용 로그인 코드 교환·가입",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "현재 기기의 Mori 로그인 세션 종료",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/providers": {
      "get": {
        "operationId": "available_providers_v1_auth_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderAvailability"
                  },
                  "title": "Response Available Providers V1 Auth Providers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "소셜 로그인 제공자",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Mori 토큰 갱신",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/{provider}/callback": {
      "get": {
        "operationId": "callback_v1_auth__provider__callback_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Provider"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{43}$",
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4096,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "302": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "소셜 제공자의 서버 콜백",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/{provider}/login": {
      "get": {
        "operationId": "login_v1_auth__provider__login_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Provider"
            }
          },
          {
            "in": "query",
            "name": "return_url",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "title": "Return Url",
              "type": "string"
            }
          },
          {
            "description": "클라이언트 verifier의 S256 값",
            "in": "query",
            "name": "code_challenge",
            "required": true,
            "schema": {
              "description": "클라이언트 verifier의 S256 값",
              "pattern": "^[A-Za-z0-9_-]{43}$",
              "title": "Code Challenge",
              "type": "string"
            }
          },
          {
            "description": "클라이언트에서 생성하고 복귀 시 확인할 난수",
            "in": "query",
            "name": "client_state",
            "required": true,
            "schema": {
              "description": "클라이언트에서 생성하고 복귀 시 확인할 난수",
              "pattern": "^[A-Za-z0-9_-]{32,128}$",
              "title": "Client State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "시스템 브라우저에서 소셜 로그인 시작",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "내 소셜 가입 정보",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}